Protect CUD items routes

+++ backend/routes/items.js
@@ -1,8 +1,11 @@
 import Router from "express-promise-router";
+import { isAdmin } from "../middleware/authMiddleware.js";
+import passport from "passport";
 import * as db from '../db/index.js'
 
 const router = new Router();
 
+// Get all items
 router.get('/', async (req, res, next) => {
   try {
     let query = 'SELECT * FROM items WHERE TRUE'; // TRUE allows for dynamic AND conditions
@@ -26,6 +29,7 @@ router.get('/', async (req, res, next) => {
   }
 });
 
+// Get an item
 router.get('/:id', async (req, res) => {
   const id = parseInt(req.params.id);
   const query = 'SELECT * FROM items WHERE id = $1';
@@ -44,7 +48,16 @@ router.get('/:id', async (req, res) => {
   }
 });
 
-router.post('/', async (req, res) => {
+// Create an item
+router.post('/', (req, res, next) => {
+  passport.authenticate('session', { session: true })(req, res, next);
+}, isAdmin, async (req, res) => {
+
+  // Check if the user is authenticated
+  if (!req.user.id) {
+    return res.status(401).json({ error: 'Unauthorized: You need to be logged in to access this order' });
+  }
+
   try {
     const newItem = req.body;
 
@@ -71,9 +84,15 @@ router.post('/', async (req, res) => {
   }
 });
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', (req, res, next) => {
+  passport.authenticate('session', { session: true })(req, res, next);
+}, isAdmin, async (req, res) => {
   const id = parseInt(req.params.id);
   const { name, type, rarity, price } = req.body;
+  // Check if the user is authenticated
+  if (!req.user.id) {
+    return res.status(401).json({ error: 'Unauthorized: You need to be logged in to access this order' });
+  }
 
   if (!name && !type && !rarity && !price) {
     return res.status(400).json({ error: 'No fields to update' });
@@ -109,8 +128,15 @@ router.put('/:id', async (req, res) => {
   }
 });
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', (req, res, next) => {
+  passport.authenticate('session', { session: true })(req, res, next);
+}, isAdmin, async (req, res) => {
   const id = parseInt(req.params.id);
+
+  // Check if the user is authenticated
+  if (!req.user.id) {
+    return res.status(401).json({ error: 'Unauthorized: You need to be logged in to access this order' });
+  }
 
   try {
     const query = 'DELETE FROM items WHERE id = $1 RETURNING *;';
